+++ packages/authhero/src/constants.ts
@@ -1,1 +1,3 @@
 export const JWKS_CACHE_TIMEOUT_IN_SECONDS = 60 * 5; // 5 minutes
+export const SILENT_AUTH_MAX_AGE = 30 * 24 * 60 * 60; // 30 days
+export const SILENT_COOKIE_NAME = "auth-token";